requirements.txt
feedparser==6.0.11
trafilatura==1.12.2
requests==2.32.3
beautifulsoup4==4.13.3
sqlalchemy==2.0.40
pymysql==1.1.1
ollama==0.4.7
python-dotenv==1.1.0
